[inductor] Consolidate codegen functions in sizevars.py into wrapper.py
Summary: Refactor the code so that wrapper codegen doesn't mix Python and C++.

[ghstack-poisoned]
