+++ torch/_inductor/graph.py
@@ -37,7 +37,7 @@ from .lowering import (
     make_fallback,
     needs_realized_inputs,
 )
-from .sizevars import CppSizeVarAllocator, SizeVarAllocator
+from .sizevars import SizeVarAllocator
 from .utils import (
     convert_shape_to_inductor,
     gather_origins,
@@ -555,7 +555,6 @@ class GraphLowering(torch.fx.Interpreter):
         if config.cpp_wrapper:
             self.check_cpp_wrapper()
             if self._can_use_cpp_wrapper:
-                self.sizevars = CppSizeVarAllocator(self._shape_env)
                 self.wrapper_code = (
                     CppAotWrapperCodeGen() if self.aot_mode else CppWrapperCodeGen()
                 )
